PARTIALLY TESTED: Make NPCs not pop into existence
NPCs aren't shown until the player fully moves onto a tile.
Fix this by showing sprites if they're one tile off-screen
(or two tiles below the bottom of the screen to allow for
the sprite's four pixels extending above the tile).

This happens to fix one cause of the trainer escape glitch
is fixed by preventing the player from opening the Start
menu in a trainer's line of sight. The order of execution
is:

1. UpdateSprites
2. AdvancePlayerSprite (almost finished with movement)
3. UpdateSprites
4. AdvancePlayerSprite (finished movement)
5. TrainerEngage
6. OverworldLoopLessDelay.displayDialogue (Start menu)
7. UpdateSprites
8. TrainerEngage

After step 4, the trainer should have been visible, but
wasn't, so the trainer isn't found in step 5. UpdateSprites
and TrainerEngage aren't called again until after step 6,
letting the Start menu appear (if Start is held). This
commit makes the trainer visible in step 3.